requirements.txt
asgiref==3.8.1
Django==5.2.9
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
inflection==0.5.1
PyYAML==6.0.1
environ
gunicorn==21.2.0
